chore(release): prepare 0.3.5
- bump Python and Rust package metadata to 0.3.5
- refresh README and docs index for streaming and multipart uploads
- centralize upload chunk validation across content= and files=
