+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "foghttp"
-version = "0.3.4"
+version = "0.3.5"
 edition = "2021"
 license = "MIT"
